[feedback] 재생 제어 로직 분리 (usePlayback, 훅)
